requirements-cloudrun.txt
-r requirements.txt
Flask>=3.0.0,<4.0
gunicorn>=23.0.0,<24.0
google-cloud-storage>=2.18.0,<4.0
